requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.6.0
yandex-cloud-ml-sdk>=0.1.0
streamlit>=1.28.0
